Added linux dev environment set up instructions
I went ahead and added the instructions to get this set up on linux. I ran from scratch and was able to get it running, all the way to watching the test GUI do its thing. This was all done on ubuntu 16.04 so I can not speak for other distro's.