Clean Up Linux Instructions
Adjusted the format of the instructions to better match the standards we are going for.  I've tested the commands on a VM and all looked good except for the eclipse one, which was adjusted.  I also added a note that the apt-get only has an older version of eclipse for some reason.